backfill/requirements.txt
googleads
google-cloud-bigquery
google-cloud-secret-manager
pandas
db-dtypes
fsspec
gcsfs
python-dateutil
pyyaml
requests
